Moved ohlc/candle series into its own module
Also support for only one `mainValue`, by getting the `closeValue` from the next day's `openValue`. The bollinger stuff is based off the `openValue` anyway.
